Make bold text render using heavier font
This is a fix to the fact that the base font does not support weights,
so bold text actually has to be a different font. I believe this was to
reduce the size of the base font for mobile usage, and could be fixed by
distributing a font which supports weights, at which point this should
be removed.